% Generated by roxygen2: do not edit by hand
\name{apply.data.cube}
\alias{apply.data.cube}
\title{Apply function over data.cube}
\usage{
apply.data.cube(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{data.cube object}

\item{MARGIN}{character or integer, dimensions by which aggregate is made}

\item{FUN}{function, by default it will apply \code{fun.aggregate} defined for each measure}

\item{...}{arguments passed to \emph{FUN}}
}
\description{
Wraps to \code{[.data.cube}.
}
\note{
When \code{FUN} argument was used, new data.cube is created with new measures.
}

