% Generated by roxygen2: do not edit by hand
\name{as.cube}
\alias{as.cube}
\title{Cast to OLAP cube}
\usage{
as.cube(x, ...)
}
\arguments{
\item{x}{R object}

\item{\dots}{arguments passed to methods}
}
\value{
\emph{cube} class object.
}
\description{
Converts arguments to \emph{cube} class. Supports \emph{list}, \emph{array} (no hierarchies), \emph{data.table}.
}
\note{
This class will be deprecated, use \code{\link{data.cube}} class instead.
}

