% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cube.R
\name{as.data.cube.data.frame}
\alias{as.data.cube.data.frame}
\title{Transform a \code{data.frame} into a \code{data.cube}.}
\usage{
\method{as.data.cube}{data.frame}(df, dim, var)
}
\arguments{
\item{df}{A \code{data.frame} to be transformed into a \code{data.cube}.}

\item{dim}{A list of unquoted expressions giving the columns of \code{df} to be used as
dimensions for the \code{data.cube}. If names are provided, they are
interpreted as names for the corresponding dimensions.}

\item{var}{A list of unquoted expressions giving the columns of \code{df} to be used as
variables for the \code{data.cube}. If names are provided, they are
interpreted as names for the corresponding variables.}
}
\value{
A \code{data.cube} resulting from the transformation of \code{df}.
}
\description{
\code{as.data.cube} transforms a \code{data.frame} into a
\code{data.cube} by interpreting its rows as observations and its
columns either as categorical descriptions of the observations (elements within
\code{data.cube} dimensions) or as numerical descriptions of the observations (values taken by
the \code{data.cube} variables).
}
