% Generated by roxygen2: do not edit by hand
\name{as.data.cube}
\alias{as.data.cube}
\title{Build fact}
\usage{
as.data.cube(x, dimensions, ...)
}
\arguments{
\item{x}{fact class object.}

\item{dimensions}{list of dimension class objects.}

\item{\dots}{ignored.}
}
\value{
fact class object.
}
\description{
Build fact
}

