% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cube.R
\name{as.data.frame.data.cube}
\alias{as.data.frame.data.cube}
\title{Transform a \code{data.cube} into a \code{data.frame}.}
\usage{
\method{as.data.frame}{data.cube}(dc, complete = FALSE)
}
\arguments{
\item{dc}{A \code{data.cube} to be transformed into a \code{data.frame}.}

\item{complete}{A logical indicating if observations for which
variables all equal zero should be presented in the output
\code{data.frame}. If not, the output \code{data.frame} hence provides
a sparse representation of the data.}
}
\value{
A \code{data.frame} resulting from the transformation of \code{dc}.
}
\description{
\code{as.data.frame} transforms a \code{data.cube} into a
\code{data.frame} by presenting observations in rows, and dimensions and
variables in columns.
}
