% Generated by roxygen2: do not edit by hand
\name{as.hierarchy}
\alias{as.hierarchy}
\alias{as.hierarchy.default}
\alias{as.hierarchy.list}
\title{Create hierarchy}
\usage{
as.hierarchy(x, ...)

\method{as.hierarchy}{default}(x, ...)

\method{as.hierarchy}{list}(x, ...)
}
\arguments{
\item{x}{list or object with a \emph{as.list} method.}

\item{\dots}{arguments passed to methods.}
}
\value{
hierarchy class object.
}
\description{
Create hierarchy
}

