% Generated by roxygen2: do not edit by hand
\name{as.level}
\alias{as.level}
\alias{as.level.data.table}
\alias{as.level.default}
\title{Create level}
\usage{
as.level(x, ...)

\method{as.level}{default}(x, id.vars = key(x), properties, ...)

\method{as.level}{data.table}(x, id.vars = key(x), properties, ...)
}
\arguments{
\item{x}{data.table or object with a \emph{as.data.table} method, build level based on that dataset.}

\item{id.vars}{character scalar of level primary key.}

\item{properties}{character vector of column names from dataset to include on that levell.}

\item{\dots}{arguments passed to methods.}
}
\value{
level class object.
}
\description{
Create level
}

