% Generated by roxygen2: do not edit by hand
\name{as.measure}
\alias{as.measure}
\title{Form measure}
\usage{
as.measure(x, label = character(), fun.aggregate = "sum", ...,
  fun.format = function(x) x)
}
\arguments{
\item{x}{character scalar column name of a measure.}

\item{label}{character scalar, a label for a measure.}

\item{fun.aggregate}{character scalar function name.}

\item{fun.format}{function to be used for formatting of a measure, such a currency.}

\item{\dots}{arguments passed to \emph{fun.aggregate}.}
}
\value{
measure class object.
}
\description{
Form measure
}

