% Generated by roxygen2: do not edit by hand
\docType{class}
\name{cube}
\alias{cube}
\title{OLAP cube class}
\format{An R6 class object.}
\usage{
cube
}
\description{
OLAP cube class
}
\details{
Generates cube class objects.
}
\note{
Consider to use \code{\link{data.cube}} class objects instead. \code{data.cube} is a set of R6 classes for cube/fact/dimension/etc and will be more actively developed than \code{cube} class.
}
\keyword{datasets}

