% Generated by roxygen2: do not edit by hand
\name{currency.format}
\alias{currency.format}
\title{Function to format numeric as currency}
\usage{
currency.format(x, currency.sym = "$", sym.align = "left", digits = 2,
  sep = ",", decimal = ".", ...)
}
\arguments{
\item{x}{numeric vector to format.}

\item{currency.sym}{character default \code{"$"}.}

\item{sym.align}{character default \code{"left"}, alternatively \code{"right"}.}

\item{digits}{precision in rounding}

\item{sep}{character separator for thousands, default \code{","}.}

\item{decimal}{character separator for decimal places, default \code{"."}.}

\item{\dots}{ignored.}
}
\value{
character vector of the same length as input formatted as defined currency.
}
\description{
Function to format numeric as currency
}

