% Generated by roxygen2: do not edit by hand
\docType{package}
\name{data.cube-package}
\alias{data.cube-package}
\title{OLAP cube data type}
\description{
Extends array for OLAP operations on multidimensional hierarchical data powered by data.table.
}
\details{
Implements OLAP cubes as data.cube object, set of data.tables objects. Aims for scalable real-time performance.
}
\author{
Jan Gorecki
}
\references{
\href{https://stackoverflow.com/questions/35472639/star-schema-normalized-dimensions-denormalized-hierarchy-level-keys}{data model inside data.cube}

\href{https://github.com/Rdatatable/data.table/wiki}{data.table}
}
\seealso{
\code{\link{as.data.cube}}, \code{\link{data.cube}}, \code{\link{[.data.cube}}, \code{\link{fact}}, \code{\link{dimension}}
}

