% Generated by roxygen2: do not edit by hand
\docType{data}
\name{data.cube}
\alias{data.cube}
\title{Data.cube class}
\format{\preformatted{Class 'R6ClassGenerator' <data.cube> object generator
  Public:
    fact: NULL
    dimensions: list
    initialize: function (fact, dimensions) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:data.cube>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "data.cube_generator"
}}
\usage{
data.cube
}
\description{
Data.cube class
}
\details{
Class stores fact class and dimension classes.
}
\keyword{datasets}

