% Generated by roxygen2: do not edit by hand
\docType{class}
\name{dimension}
\alias{dimension}
\title{Dimension class}
\format{An R6 class object.}
\usage{
dimension
}
\description{
Dimension class
}
\details{
Class stores set of hierarchies. Initialized with hierarchies list definition. Also stores mapping from primary key to any level, to use snowflake
}
\keyword{datasets}

