% Generated by roxygen2: do not edit by hand
\docType{class}
\name{fact}
\alias{fact}
\title{Fact class}
\format{An R6 class object.}
\usage{
fact
}
\description{
Fact class
}
\details{
Class stores fact table as local \code{data.table} or remote \code{big.data.table}. Measures can be provided manually to \code{measures} argument, useful for custom aggregate function per measure.
}
\seealso{
\code{\link{as.fact}}, \code{\link{measure}}, \code{\link{dimension}}, \code{\link{data.cube}}
}
\keyword{datasets}

