% Generated by roxygen2: do not edit by hand
\docType{class}
\name{level}
\alias{level}
\title{Level class}
\format{An R6 class object.}
\usage{
level
}
\description{
Level class
}
\details{
Class stores lower grain of dimension attributes. Initialized with key column and properties columns - all depndent on the key.
}
\seealso{
\code{\link{as.level}}, \code{\link{hierarchy}}, \code{\link{dimension}}, \code{\link{data.cube}}
}
\keyword{datasets}

