% Generated by roxygen2: do not edit by hand
\docType{data}
\name{measure}
\alias{measure}
\title{Measure class}
\format{\preformatted{Class 'R6ClassGenerator' <measure> object generator
  Public:
    var: 
    fun.aggregate: 
    dots: list
    label: 
    initialize: function (x, label = character(), fun.aggregate = "sum", ...) 
    format: function () 
    print: function () 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:data.cube>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "measure_generator"
}}
\usage{
measure
}
\description{
Measure class
}
\details{
Class stores variable name from fact table, the function to use against variable. Initialized with character scalar variable name, optional label, function to use on aggregates and it's custom arguments. Method `format` is provided to produce clean expressions.
}
\keyword{datasets}

