% Generated by roxygen2: do not edit by hand
\name{process.dim}
\alias{process.dim}
\title{Process dimension}
\usage{
process.dim(dim, x)
}
\arguments{
\item{dim}{character scalar name of dimension}

\item{x}{list value for element *dim*}
}
\description{
Helps in processing dimension data for results of *dimnames.array* function. It also check *key* and it's uniqueness.
}

