% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cube.R
\name{property.data.cube}
\alias{property.data.cube}
\alias{dim.names.data.cube}
\alias{elm.names.data.cube}
\alias{var.names.data.cube}
\alias{dim.nb.data.cube}
\alias{elm.nb.data.cube}
\alias{var.nb.data.cube}
\title{Get basic properties.}
\usage{
\method{dim.names}{data.cube}(dc)

\method{elm.names}{data.cube}(dc, ...)

\method{var.names}{data.cube}(dc)

\method{dim.nb}{data.cube}(dc)

\method{elm.nb}{data.cube}(dc, ...)

\method{var.nb}{data.cube}(dc)
}
\arguments{
\item{dc}{A \code{data.cube}.}

\item{...}{Optional. Names of the dimensions of interest.}
}
\value{
A character vector giving the names of the dimensions of
\code{dc}.

A named list of character vectors giving the names of the
elements in the dimensions of \code{dc}.

A character vector giving the names of the variables of
\code{dc}.

An integer giving the number of dimensions of \code{dc}.

A named list of integers giving the number of elements in the
dimensions of \code{dc}.

An integer giving the number of variables of \code{dc}.
}
\description{
A collection of functions to get basic properties of a \code{data.cube}.
}
