% Generated by roxygen2: do not edit by hand
\name{rollup}
\alias{rollup}
\title{Rollup generic method}
\usage{
rollup(x, ...)
}
\arguments{
\item{x}{R object}

\item{\dots}{arguments passed to methods}
}
\value{
Same class object after \emph{rollup} aggregation applied.
}
\description{
Calculate rollup aggregations on provided attributes.
}

