% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cube.R
\name{select.dim.data.cube}
\alias{select.dim.data.cube}
\title{Select dimensions of interest.}
\usage{
\method{select.dim}{data.cube}(dc, ...)
}
\arguments{
\item{dc}{A datacube.}

\item{...}{A list of unquoted expressions giving the dimensions to select.}
}
\value{
The resulting datacube.
}
\description{
\code{select.dim} selects a subset of dimensions within the \code{data.cube} by aggregating together the elements of each unselected dimension.
This consists in replacing the datacube by one of its marginal dataplane.
}
