% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cube.R
\name{select.elm.data.cube}
\alias{select.elm.data.cube}
\alias{remove.elm.data.cube}
\title{Select or remove elements within a given dimension.}
\usage{
\method{select.elm}{data.cube}(dc, dim, elm = NULL, top.nb = NULL,
  bot.nb = NULL, var = NULL, suppress = FALSE)

\method{remove.elm}{data.cube}(dc, dim, elm = NULL, top = NULL,
  bot = NULL, var = NULL, suppress = FALSE)
}
\arguments{
\item{dc}{A datacube.}

\item{dim}{An unquoted expression giving the dimension whose elements are selected or removed.}

\item{elm}{Optional. A character vector giving the names of the elements to select or to remove.}

\item{top.nb}{bot.nb Optional. An integer giving the number of elements to select or to remove among the ones with the highest or the lowest values of the specified variable (see \code{var}).}

\item{var}{Optional. Can be specified if either \code{top.nb} or \code{bot.nb} is specified. An unquoted expression giving the variable to consider when selecting or removing the top or the bottom elements. If not specified, the datacube's first variable is taken. Can also be \code{model}, \code{ratio}, or \code{deviation} if a marginal model has been computed.}

\item{suppress}{A logical indicating if the corresponding observations should be fully suppressed (recomputing the datacube's marginals) or simply hidden (no longer shown when processing the datacube).}
}
\value{
The resulting datacube.
}
\description{
\code{select.elm} and \code{remove.elm} respectively select and remove a subset of elements within a particular dimensions.
This consists in suppressing all observations that corresponds to the unselected or removed elements.
}
