% Generated by roxygen2: do not edit by hand
\name{[.cube}
\alias{[.cube}
\title{Query cube}
\usage{
\method{[}{cube}(x, i, j, by, drop = TRUE)
}
\arguments{
\item{x}{cube object}

\item{i}{list of values used to slice and dice on cube}

\item{j}{expression to evaluate on fact}

\item{by}{expression/character vector to aggregate measures accroding to *j* argument.}

\item{drop}{logical, default TRUE, drop dimensions, logically the same as *drop* argument in `[.array`.}
}
\value{
Cube class object
}
\description{
Query cube
}

