% Generated by roxygen2: do not edit by hand
\name{[.data.cube}
\alias{[.data.cube}
\title{Subset data.cube}
\usage{
\method{[}{data.cube}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{data.cube object}

\item{...}{values to subset on corresponding dimensions, when wrapping in list it will refer to dimension hierarchies}

\item{drop}{logical, default TRUE, drop redundant dimensions, same as *drop* argument in \code{[.array}.}
}
\value{
data.cube class object
}
\description{
Subset data.cube
}

