% Generated by roxygen2: do not edit by hand
\name{[[.cube}
\alias{[[.cube}
\title{Extract cube}
\usage{
\method{[[}{cube}(x, i, j, by)
}
\arguments{
\item{x}{cube object}

\item{i}{list of values used to slice and dice on cube}

\item{j}{expression to evaluate on fact}

\item{by}{expression/character vector to aggregate measures accroding to *j* argument.}
}
\value{
Cube class object
}
\description{
Extract cube
}
\note{
Consider to use \code{\link{data.cube}} class objects instead. \code{data.cube} is a set of R6 classes for cube/fact/dimension/etc and will be more actively developed than \code{cube} class.
}

