% Generated by roxygen2: do not edit by hand
\name{as.cube}
\alias{as.cube}
\title{Cast to OLAP cube}
\usage{
as.cube(x, ...)
}
\arguments{
\item{x}{R object}

\item{\dots}{arguments passed to methods}
}
\value{
*cube* class object.
}
\description{
Converts arguments to *cube* class. Supports *list*, *array* (no hierarchies), *data.table*.
}
\note{
Consider to use \code{\link{data.cube}} class objects instead. \code{data.cube} is a set of R6 classes for cube/fact/dimension/etc and will be more actively developed than \code{cube} class.
}

