% Generated by roxygen2: do not edit by hand
\name{as.data.cube}
\alias{as.data.cube}
\alias{as.data.cube.array}
\alias{as.data.cube.cube}
\alias{as.data.cube.data.table}
\alias{as.data.cube.default}
\alias{as.data.cube.fact}
\alias{as.data.cube.list}
\alias{as.data.cube.matrix}
\title{Build cube}
\usage{
as.data.cube(x, ...)

\method{as.data.cube}{default}(x, ...)

\method{as.data.cube}{matrix}(x, na.rm = TRUE, ...)

\method{as.data.cube}{array}(x, na.rm = TRUE, ...)

\method{as.data.cube}{fact}(x, dimensions = list(), ...)

\method{as.data.cube}{list}(x, ...)

\method{as.data.cube}{data.table}(x, id.vars = key(x), measure.vars,
  fun.aggregate = sum, dims = id.vars, hierarchies = NULL, ...,
  dimensions, measures = NULL)

\method{as.data.cube}{cube}(x, hierarchies = NULL, ...)
}
\arguments{
\item{x}{R object.}

\item{na.rm}{logical, default TRUE, when FALSE the cube would store cross product of all dimension grain keys!}

\item{dimensions}{list of \link{dimension} class objects.}

\item{id.vars}{characater vector of foreign key columns.}

\item{measure.vars}{characater vector of column names of metrics.}

\item{fun.aggregate}{function, default to \emph{sum}.}

\item{dims}{character vector of dimension names}

\item{hierarchies}{list of hierarchies nested in list of dimensions passed to \link{as.dimension}.}

\item{measures}{list of \link{measure} class objects passed to \link{as.fact}.}

\item{\dots}{arguments passed to methods.}
}
\value{
data.cube class object.
}
\description{
Build cube
}
\examples{
library(data.table)
set.seed(1L)
dt = CJ(color = c("green","yellow","red"),
        year = 2011:2015,
        status = c("active","inactive","archived","removed"))[sample(30)]
dt[, "value" := sample(4:7/2, nrow(dt), TRUE)]

# from data.table
dc = as.data.cube(
    x = dt, id.vars = c("color","year","status"),
    measure.vars = "value",
    hierarchies = sapply(c("color","year","status"),
                         function(x) list(setNames(list(character()), x)),
                         simplify=FALSE)
)
str(dc)

# multidimensional hierarchical data from fact and dimensions
X = populate_star(N = 1e3)
sales = X$fact$sales
time = X$dims$time
geography = X$dims$geography
# define hierarchies
time.hierarchies = list( # 2 hierarchies in time dimension
    "monthly" = list(
        "time_year" = character(),
        "time_quarter" = c("time_quarter_name"),
        "time_month" = c("time_month_name"),
        "time_date" = c("time_month","time_quarter","time_year")
    ),
    "weekly" = list(
        "time_year" = character(),
        "time_week" = character(),
        "time_date" = c("time_week","time_year")
    )
)
geog.hierarchies = list( # 1 hierarchy in geography dimension
    list(
        "geog_region_name" = character(),
        "geog_division_name" = c("geog_region_name"),
        "geog_abb" = c("geog_name","geog_division_name","geog_region_name")
    )
)
# create dimensions
dims = list(
    time = as.dimension(x = time,
                        id.vars = "time_date",
                        hierarchies = time.hierarchies),
    geography = as.dimension(x = geography,
                             id.vars = "geog_abb",
                             hierarchies = geog.hierarchies)
)
# create fact
ff = as.fact(
    x = sales,
    id.vars = c("geog_abb","time_date"),
    measure.vars = c("amount","value"),
    fun.aggregate = sum,
    na.rm = TRUE
)
# create data.cube
dc = as.data.cube(ff, dims)
str(dc)
}
\seealso{
\code{\link{data.cube}}, \code{\link{fact}}, \code{\link{dimension}}
}

