% Generated by roxygen2: do not edit by hand
\name{as.data.cube}
\alias{as.data.cube}
\alias{as.data.cube.array}
\alias{as.data.cube.default}
\alias{as.data.cube.fact}
\title{Build cube}
\usage{
as.data.cube(x, ...)

\method{as.data.cube}{default}(x, ...)

\method{as.data.cube}{array}(x, na.rm = TRUE, ...)

\method{as.data.cube}{fact}(x, dimensions, ...)
}
\arguments{
\item{x}{R object.}

\item{na.rm}{logical, default TRUE, when FALSE the cube would store cross product of all dimension grain keys!}

\item{dimensions}{list of dimension class objects.}

\item{\dots}{arguments passed to methods.}
}
\value{
data.cube class object.
}
\description{
Build cube
}

