% Generated by roxygen2: do not edit by hand
\name{as.data.cube}
\alias{as.data.cube}
\alias{as.data.cube.array}
\alias{as.data.cube.cube}
\alias{as.data.cube.data.table}
\alias{as.data.cube.default}
\alias{as.data.cube.fact}
\alias{as.data.cube.list}
\alias{as.data.cube.matrix}
\title{Build cube}
\usage{
as.data.cube(x, ...)

\method{as.data.cube}{default}(x, ...)

\method{as.data.cube}{matrix}(x, na.rm = TRUE, ...)

\method{as.data.cube}{array}(x, na.rm = TRUE, ...)

\method{as.data.cube}{fact}(x, dimensions = list(), ...)

\method{as.data.cube}{list}(x, ...)

\method{as.data.cube}{data.table}(x, id.vars = key(x), measure.vars,
  fun.aggregate = "sum", dims = id.vars, hierarchies = NULL, ...,
  dimensions, measures)

\method{as.data.cube}{cube}(x, hierarchies = NULL, ...)
}
\arguments{
\item{x}{R object.}

\item{na.rm}{logical, default TRUE, when FALSE the cube would store cross product of all dimension grain keys!}

\item{dimensions}{list of \link{dimension} class objects.}

\item{id.vars}{characater vector of foreign key columns.}

\item{measure.vars}{characater vector of column names of metrics.}

\item{fun.aggregate}{character function name, default to \emph{sum}.}

\item{dims}{character vector of dimension names}

\item{hierarchies}{list of hierarchies nested in list of dimensions passed to \link{as.dimension}.}

\item{measures}{list of \link{measure} class objects passed to \link{as.fact}.}

\item{\dots}{arguments passed to methods.}
}
\value{
data.cube class object.
}
\description{
Build cube
}

