% Generated by roxygen2: do not edit by hand
\name{as.dimension}
\alias{as.dimension}
\title{Build dimension}
\usage{
as.dimension(x, key = key(x), hierarchies)
}
\arguments{
\item{x}{data.table build dimension based on that dataset.}

\item{key}{character scalar of dimension's primary key column.}

\item{hierarchies}{list of hierarchies and levels within dimensions.}
}
\value{
dimension class object.
}
\description{
Build dimension
}

