% Generated by roxygen2: do not edit by hand
\name{as.dimension}
\alias{as.dimension}
\alias{as.dimension.data.table}
\alias{as.dimension.default}
\title{Build dimension}
\usage{
as.dimension(x, ...)

\method{as.dimension}{default}(x, id.vars, hierarchies = NULL, ...)

\method{as.dimension}{data.table}(x, id.vars = key(x), hierarchies = NULL,
  ...)
}
\arguments{
\item{x}{data.table or object with a \emph{as.data.table} method, build dimension based on that dataset.}

\item{id.vars}{character scalar of dimension primary key.}

\item{hierarchies}{list of hierarchies of levels and its attributes.}

\item{\dots}{arguments passed to methods.}
}
\value{
dimension class object.
}
\description{
Build dimension
}

