% Generated by roxygen2: do not edit by hand
\name{as.fact}
\alias{as.fact}
\alias{as.fact.data.table}
\alias{as.fact.default}
\title{Build fact}
\usage{
as.fact(x, ...)

\method{as.fact}{default}(x, id.vars = character(),
  measure.vars = character(), fun.aggregate = "sum", ..., measures)

\method{as.fact}{data.table}(x, id.vars = character(),
  measure.vars = character(), fun.aggregate = "sum", ..., measures)
}
\arguments{
\item{x}{data.table build dimension based on that dataset.}

\item{id.vars}{character vector of all dimension's foreign keys.}

\item{measure.vars}{character vector, column names of measures.}

\item{fun.aggregate}{default "sum".}

\item{measures}{list of measures class objects, useful if various measures needs to have different `fun.aggregate`.}

\item{\dots}{arguments to fun.aggregate.}
}
\value{
fact class object.
}
\description{
Build fact
}

