% Generated by roxygen2: do not edit by hand
\name{as.hierarchy}
\alias{as.hierarchy}
\alias{as.hierarchy.default}
\alias{as.hierarchy.list}
\title{Create hierarchy}
\usage{
as.hierarchy(x, ...)

\method{as.hierarchy}{default}(x, ...)

\method{as.hierarchy}{list}(x, ...)
}
\arguments{
\item{x}{list which keeps named character vectors (often length 0L) of level key and dependent attributes not listed yet on lower level key.}

\item{\dots}{arguments passed to methods.}
}
\value{
hierarchy class object.
}
\description{
Create hierarchy
}
\details{
You generally won't need to use that function directly, but use \code{\link{as.dimension}} that takes a list of (possibly named) hierarchies inputs.
}
\examples{
time.calendar = as.hierarchy(list(
  year = character(),
  quarter = character(),
  month = character(),
  date = c("year","quarter","month")
))
str(time.calendar)
}
\seealso{
\code{\link{hierarchy}}, \code{\link{level}}, \code{\link{dimension}}, \code{\link{data.cube}}
}

