% Generated by roxygen2: do not edit by hand
\name{capply}
\alias{capply}
\title{apply over cube dimensions}
\usage{
capply(x, MARGIN, FUN, ...)
}
\arguments{
\item{x}{cube object}

\item{MARGIN}{character or list}

\item{FUN}{function}

\item{...}{arguments passed to *FUN*}
}
\description{
apply over cube dimensions
}

