% Generated by roxygen2: do not edit by hand
\name{capply}
\alias{capply}
\title{Apply function on measures while aggregate on cube dimensions}
\usage{
capply(x, MARGIN, FUN, ...)
}
\arguments{
\item{x}{cube object}

\item{MARGIN}{character or list}

\item{FUN}{function}

\item{...}{arguments passed to *FUN*}
}
\description{
Wrapper around `[[.cube` and `j`, `by` arg.
}
\note{
Consider to use \code{\link{data.cube}} class objects instead. \code{data.cube} is a set of R6 classes for cube/fact/dimension/etc and will be more actively developed than \code{cube} class.
}

