% Generated by roxygen2: do not edit by hand
\docType{class}
\name{data.cube}
\alias{data.cube}
\title{data.cube class}
\format{An R6 class object.}
\usage{
data.cube
}
\description{
data.cube class
}
\details{
Class stores fact class and dimension classes. R6 class preview with \code{ls.str()}. For all tables in the model of cube use \code{str()}.
}
\seealso{
\code{\link{data.cube-package}}, \code{\link{as.data.cube}}, \code{\link{[.data.cube}}, \code{\link{fact}}, \code{\link{dimension}}
}
\keyword{datasets}

