% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cube.R
\docType{package}
\name{data.cube}
\alias{data.cube}
\alias{data.cube-package}
\title{Easy Processing of Multidimensional Data.}
\description{
\code{data.cube} is an R package for the exploration of multidimensional
datasets and for the detection of statistical outliers within. It is
mainly a tool for data exploration, allowing to have a first glance at
it and to formulate research hypotheses to be later tested.

The package defines a new data structure called data.cube that can be
fed with a classical \code{data.frame} encoding a list of numeric
variables described according to several categorical dimensions. For
example, in the case of Twitter data, it can be the number of tweets
(numeric variable) that have been published by a given user (first
dimension) about a given topic (second dimension) at a given date (third
dimension). The input data.frame hence takes the form of a list of
quadruplets (user, topic, date, number of tweets).

Statistical outliers can then be identified among the observations by
first selecting some dimensions of interest, that is by subsetting or
by aggregating the input dimensions. If needed, observations can also be
normalised according to the marginal values along the selected
dimensions, thus comparing the observed value to an expected value
obtained by the uniform redistribution of the selected marginal values.
Different statistical tests can then be chosen to measure the deviation
between the observed and the expected values. The package finally allows
to retrieve a list of positive outliers, that is observations that are
significantly higher than expected.

Note that the current implementation is optimised for sparse data. 

This package has been developed by researchers of the Complex Networks
team, within the Computer Science Laboratory of Paris 6 (LIP6), for the
ODYCCEUS project, founded by the European Commission FETPROACT 2016-2017
program under grant 732942.

Links:
\itemize{
\item Complex Networks team: \url{http://www.complexnetworks.fr/}
\item LIP6: \url{https://www.lip6.fr/}
\item ODYCCEUS project: \url{https://www.odycceus.eu/}
}

Contact:
\itemize{
\item Robin Lamarche-Perrin: \email{Robin.Lamarche-Perrin@lip6.fr}

See also my webpage:
\url{https://www-complexnetworks.lip6.fr/~lamarche/}
}

List of main collaborators:
\itemize{
\item Robin Lamarche-Perrin (CNRS, ISC-PIF, LIP6)
\item Audrey Wilmet (UPMC, LIP6)
\item Léonard Panichi (UPMC, LIP6)
}

Copyright 2018 © Robin Lamarche-Perrin

\code{data.cube} is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by the
Free Software Foundation, either version 3 of the License, or (at your
option) any later version. It is distributed in the hope that it will be
useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
Public License for more details. You should have received a copy of the
GNU General Public License along with this program. If not, see
\url{http://www.gnu.org/licenses/}.
}
