% Generated by roxygen2: do not edit by hand
\docType{data}
\name{dimension}
\alias{dimension}
\title{Dimension class}
\format{\preformatted{Class 'R6ClassGenerator' <dimension> object generator
  Public:
    key: 
    hierarchies: list
    data: NULL
    initialize: function (x, key = key(x), hierarchies) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:data.cube>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "dimension_generator"
}}
\usage{
dimension
}
\description{
Dimension class
}
\details{
Class stores set of hierarchies. Initialized with hierarchies list definition. Also stores mapping from primary key to any level, to use snowflake
}
\keyword{datasets}

