% Generated by roxygen2: do not edit by hand
\docType{data}
\name{fact}
\alias{fact}
\title{Fact class}
\format{\preformatted{Class 'R6ClassGenerator' <fact> object generator
  Public:
    local: 
    id.vars: 
    measure.vars: 
    measures: list
    data: NULL
    initialize: function (x, id.vars = character(), measure.vars = character(), 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:data.cube>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "fact_generator"
}}
\usage{
fact
}
\description{
Fact class
}
\details{
Class stores fact table as local data.table or remote big.data.table. Initialized with data.table or list of R nodes connections. Measures can be provided manually to `measures` argument, useful for custom aggregate function per measure.
}
\keyword{datasets}

