% Generated by roxygen2: do not edit by hand
\docType{class}
\name{fact}
\alias{fact}
\title{Fact class}
\format{An R6 class object.}
\usage{
fact
}
\description{
Fact class
}
\details{
Class stores fact table as local data.table or remote big.data.table. Initialized with data.table or list of R nodes connections. Measures can be provided manually to `measures` argument, useful for custom aggregate function per measure.
}
\keyword{datasets}

