% Generated by roxygen2: do not edit by hand
\docType{data}
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchy class}
\format{\preformatted{Class 'R6ClassGenerator' <hierarchy> object generator
  Public:
    key: 
    levels: list
    initialize: function (x, key = key(x), levels) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:data.cube>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "hierarchy_generator"
}}
\usage{
hierarchy
}
\description{
Hierarchy class
}
\details{
Class stores set of dimension levels into hierarchy. Initialized with key column and levels list.
}
\keyword{datasets}

