% Generated by roxygen2: do not edit by hand
\docType{class}
\name{hierarchy}
\alias{hierarchy}
\title{Hierarchy class}
\format{An R6 class object.}
\usage{
hierarchy
}
\description{
Hierarchy class
}
\details{
Class stores set of dimension levels into hierarchy. Currently just keeps a list.
}
\keyword{datasets}

