% Generated by roxygen2: do not edit by hand
\docType{data}
\name{level}
\alias{level}
\title{Level class}
\format{\preformatted{Class 'R6ClassGenerator' <level> object generator
  Public:
    key: 
    properties: 
    parents: 
    data: NULL
    initialize: function (x, key = key(x), properties) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:data.cube>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "level_generator"
}}
\usage{
level
}
\description{
Level class
}
\details{
Class stores lower grain of dimension attributes. Initialized with key column and propoerties columns - all depndent on the key.
}
\keyword{datasets}

