% Generated by roxygen2: do not edit by hand
\docType{class}
\name{level}
\alias{level}
\title{Level class}
\format{An R6 class object.}
\usage{
level
}
\description{
Level class
}
\details{
Class stores lower grain of dimension attributes. Initialized with key column and propoerties columns - all depndent on the key.
}
\keyword{datasets}

