% Generated by roxygen2: do not edit by hand
\name{rollup}
\alias{rollup}
\title{Rollup generic method}
\usage{
rollup(x, ...)
}
\arguments{
\item{x}{R object}

\item{\dots}{arguments passed to methods}
}
\value{
*cube* class object. Use `format` on results to print table and format measures.
}
\description{
Converts arguments to *cube* class by reducing dimensions, sub aggregates are provided as additional NA dimension value..
}

