% Generated by roxygen2: do not edit by hand
\name{[.cube}
\alias{[.cube}
\title{Subset cube}
\usage{
\method{[}{cube}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{cube object}

\item{...}{values to subset on corresponding dimensions, when wrapping in list it will refer to dimension hierarchy}

\item{drop}{logical, default TRUE, drop dimensions same as *drop* argument in `[.array`.}
}
\value{
Cube class object
}
\description{
Subset cube
}
\note{
Consider to use \code{\link{data.cube}} class objects instead. \code{data.cube} is a set of R6 classes for cube/fact/dimension/etc and will be more actively developed than \code{cube} class.
}

