% Generated by roxygen2: do not edit by hand
\name{[.cube}
\alias{[.cube}
\title{apply over cube dimensions
#}
\usage{
\method{[}{cube}(x, i, j, by, drop = TRUE)
}
\arguments{
\item{x}{cube object
#}

\item{drop}{logical, default TRUE, drop dimensions same as *drop* argument in `[.array`.}

\item{MARGIN}{character or list
#}

\item{FUN}{function
#}

\item{...}{arguments passed to *FUN*}

\item{x}{cube object}

\item{...}{values to subset on corresponding dimensions, when wrapping in list it will refer to dimension hierarchy}
}
\value{
Cube class object
}
\description{
#
}

