% Generated by roxygen2: do not edit by hand
\name{[.cube}
\alias{[.cube}
\title{subset cube}
\usage{
\method{[}{cube}(x, ..., drop = FALSE)
}
\arguments{
\item{x}{cube object}

\item{...}{values to subset on corresponding dimensions, when wrapping in list it will refer to dimension hierarchy}

\item{drop}{logical, default may switch to TRUE while implemented.}
}
\value{
When *drop* arg is TRUE then 1-cardinality dimensions will be removed, also dimension members not linked to fact will be removed.
}
\description{
subset cube
}

