% Generated by roxygen2: do not edit by hand
\name{[.cube}
\alias{[.cube}
\title{subset cube}
\usage{
\method{[}{cube}(x, ..., drop = TRUE)
}
\arguments{
\item{x}{cube object}

\item{...}{values to subset on corresponding dimensions, when wrapping in list it will refer to dimension hierarchy}

\item{drop}{logical, default TRUE, drop dimensions same as *drop* argument in `[.array`.}
}
\value{
Cube class object
}
\description{
subset cube
}

